chatbot/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
qdrant-client==1.7.0
sentence-transformers==3.0.1
google-generativeai==0.3.0
langdetect==1.0.9
transformers==4.35.2
# PyTorch with CUDA support - install with GPU index URL for best compatibility:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CPU-only, use: pip install torch torchvision torchaudio
torch==2.9.0
PyMuPDF==1.23.8
python-docx==0.8.11
numpy==1.26.1
scikit-learn==1.3.2
requests==2.31.0
pymongo==4.6.1
dnspython==2.4.2

# Vietnamese NLP for hybrid search (BM25)
underthesea==6.7.0
rank-bm25==0.2.2
